# docs/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "GrownupsVet API",
    "description" : "Contrato generado desde código para acceso, perfil, sesiones, mascotas, recuperación, personal con invitaciones y disponibilidad veterinaria. El correo requiere configuración SMTP. La URL de activación del portal contiene un placeholder configurable. Reservas, consulta automática de COMVEZCOL y plantillas Thymeleaf quedan fuera de este incremento.",
    "version" : "0.6.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/veterinarians/{veterinarianId}/availability-slot-batches" : {
      "post" : {
        "tags" : [ "Disponibilidad veterinaria" ],
        "summary" : "Generar un lote de turnos",
        "description" : "Genera filas concretas para un rango inclusivo máximo de 31 días y no más de 1000 turnos. 24:00 solo se admite como fin exclusivo.",
        "operationId" : "createVeterinarianAvailabilitySlotBatch",
        "parameters" : [ {
          "name" : "veterinarianId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateVeterinarianAvailabilitySlotBatchRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Rango o generación inválida",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Sesión ausente o inválida",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "El rol no permite crear turnos",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Veterinario no encontrado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "Un turno solicitado ya existe",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "415" : {
            "description" : "Tipo de contenido no admitido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Fallo interno sin información sensible",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "201" : {
            "description" : "Turnos creados de forma atómica",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreateVeterinarianAvailabilitySlotBatchResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiErrorResponseDTO" : {
        "type" : "object",
        "description" : "Respuesta de error RFC 9457, con código funcional y errores por campo.",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "Referencia URI al tipo de problema.",
            "example" : "about:blank"
          },
          "title" : {
            "type" : "string",
            "description" : "Resumen del problema.",
            "example" : "Solicitud incorrecta"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Estado HTTP de la respuesta.",
            "example" : 400,
            "maximum" : 599,
            "minimum" : 400
          },
          "detail" : {
            "type" : "string",
            "description" : "Explicación de lo ocurrido en esta solicitud.",
            "example" : "Revisa los campos indicados."
          },
          "instance" : {
            "type" : "string",
            "format" : "uri-reference",
            "description" : "Referencia URI opcional que identifica esta ocurrencia del problema.",
            "example" : "urn:uuid:7d667530-867b-4c78-a458-0904fb82d574"
          },
          "errorCode" : {
            "type" : "string",
            "description" : "Código funcional estable; el frontend no debe comparar los mensajes.",
            "example" : "INVALID_REQUEST"
          },
          "fieldErrors" : {
            "type" : "array",
            "description" : "Errores por campo. Es una lista vacía cuando el error no corresponde a campos específicos.",
            "items" : {
              "$ref" : "#/components/schemas/FieldValidationErrorResponseDTO"
            }
          }
        },
        "required" : [ "detail", "errorCode", "fieldErrors", "status", "title", "type" ]
      },
      "FieldValidationErrorResponseDTO" : {
        "type" : "object",
        "description" : "Error asociado a un campo de la solicitud. Nunca contiene el valor recibido.",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "Nombre del campo en el contrato JSON.",
            "example" : "email"
          },
          "code" : {
            "type" : "string",
            "description" : "Código estable que identifica el error del campo.",
            "example" : "TYPE_MISMATCH"
          },
          "message" : {
            "type" : "string",
            "description" : "Explicación legible para la persona usuaria.",
            "example" : "Debe enviarse como texto."
          }
        },
        "required" : [ "code", "field", "message" ]
      },
...
      "VeterinarianAvailabilitySlotResponseDTO" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "veterinarianId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PUBLISHED", "BLOCKED" ]
          },
          "version" : {
            "type" : "integer",
            "format" : "int64",
            "minimum" : 0
          },
          "timeZone" : {
            "type" : "string",
            "enum" : [ "America/Bogota" ]
          }
        },
        "required" : [ "endsAt", "id", "startsAt", "status", "timeZone", "version", "veterinarianId" ]
      },
...
      "CreateVeterinarianAvailabilitySlotBatchRequestDTO" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "startDate" : {
            "type" : "string",
            "format" : "date"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date"
          },
          "daysOfWeek" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int32",
              "maximum" : 7,
              "minimum" : 1
            },
            "maxItems" : 7,
            "minItems" : 1
          },
          "dailyStartTime" : {
            "type" : "string",
            "pattern" : "^(?:[01][0-9]|2[0-3]):(?:00|30)$"
          },
          "dailyEndTime" : {
            "type" : "string",
            "example" : "24:00",
            "pattern" : "^(?:(?:[01][0-9]|2[0-3]):(?:00|30)|24:00)$"
          }
        },
        "required" : [ "dailyEndTime", "dailyStartTime", "daysOfWeek", "endDate", "startDate" ]
      },
      "CreateVeterinarianAvailabilitySlotBatchResponseDTO" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "createdCount" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 1000,
            "minimum" : 1
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/VeterinarianAvailabilitySlotResponseDTO"
            },
            "maxItems" : 1000,
            "minItems" : 1
          },
          "timeZone" : {
            "type" : "string",
            "enum" : [ "America/Bogota" ]
          }
        },
        "required" : [ "createdCount", "items", "timeZone" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "JWT RS256 supplied in the Authorization header. Tokens expire 24 hours after login.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}